requirements.txt
opencv-python==4.12.0.88
streamlit==1.47.1
numpy==2.2.6
pillow==11.3.0
streamlit-webrtc==0.63.3
aioice==0.10.1
aiortc==1.13.0
av==14.4.0
pylibsrtp==0.12.0
pyee==13.0.0
cryptography==45.0.5